Remove items endpoint and move batched items call to bib fetch endpoint
