+++ pages/api/bib/[id].tsx
@@ -1,6 +1,6 @@
 import type { NextApiRequest, NextApiResponse } from "next"
-import { fetchBib } from "../../../../src/server/api/bib"
-import { PATHS, BASE_URL } from "../../../../src/config/constants"
+import { fetchBib } from "../../../src/server/api/bib"
+import { PATHS, BASE_URL } from "../../../src/config/constants"
 
 /**
  * Default API route handler for Bib page
