+++ src/server/api/bib.ts
@@ -8,10 +8,12 @@ import nyplApiClient from "../nyplApiClient"
 import {
   DISCOVERY_API_NAME,
   DISCOVERY_API_SEARCH_ROUTE,
+  ITEM_VIEW_ALL_BATCH_SIZE,
   SHEP_HTTP_TIMEOUT,
 } from "../../config/constants"
 import { appConfig } from "../../config/config"
 import logger from "../../../logger"
+import type { DiscoveryItemResult } from "../../types/itemTypes"
 
 export async function fetchBib(
   id: string,
@@ -91,6 +93,13 @@ export async function fetchBib(
         status: 404,
       }
     }
+    // Populate bib's items with all the items if View All is enabled
+    if (bibQuery?.view_all_items) {
+      discoveryBibResult.items = await fetchAllBibItems(
+        bibQuery,
+        discoveryBibResult.numItemsMatched
+      )
+    }
 
     return {
       discoveryBibResult,
@@ -126,4 +135,37 @@ async function fetchBibSubjectHeadings(bibId: string) {
   } finally {
     clearTimeout(timeoutId)
   }
+}
+
+async function fetchAllBibItems(
+  bibQuery: BibQueryParams,
+  numItems: number,
+  batchSize = ITEM_VIEW_ALL_BATCH_SIZE
+): Promise<DiscoveryItemResult[]> {
+  const items: DiscoveryItemResult[] = []
+  const client = await nyplApiClient({ apiName: DISCOVERY_API_NAME })
+
+  for (
+    let batchNum = 1;
+    batchNum <= Math.ceil(numItems / batchSize);
+    batchNum++
+  ) {
+    const pageQueryString = getBibQueryString(
+      {
+        ...bibQuery,
+        item_page: batchNum,
+      },
+      false,
+      true
+    )
+    const bibPage = await client.get(
+      `${DISCOVERY_API_SEARCH_ROUTE}/${bibQuery.id}${pageQueryString}`
+    )
+    if (bibPage?.items?.length) {
+      items.push(...bibPage.items)
+    } else {
+      throw new Error("There was en error fetching items in one of the batches")
+    }
+  }
+  return items
 }
