+++ src/types/itemTypes.ts
@@ -1,9 +1,3 @@
-/* eslint-disable @typescript-eslint/naming-convention */
-
-// Item structure coming from the Search Results API response
-import { AnnotatedMarc } from "./bibDetailsTypes"
-import type { DiscoveryBibResult } from "./bibTypes"
-
 export interface DiscoveryItemResult {
   uri?: string
   idNyplSourceId?: ItemSourceID
@@ -39,12 +33,6 @@ export interface ItemMetadata {
   barcode?: string
   callNumber?: string
   bibId?: string
-}
-
-export interface ItemsResponse {
-  items: DiscoveryItemResult[]
-  discoveryBibResult: DiscoveryBibResult
-  status: 200 | 400
 }
 
 /* eslint-disable @typescript-eslint/naming-convention */
